Split ConnectionDetails into focused components
